Planning, Migrations & Helpers
Planning has been added, and we can now update the agenda and get the agenda via endpoints.

Updated the migrations, this is so we have better priority. This is when you have a migration that depends on another migration having ran. Using a prefix of 000__ you can define which runs first

Changed helpers, helpers now follow the laravel naming convention!