Attendance Status
Added a seperate model for attendance status. With this its more customizable what statuses there are for attendance
Changed testing with this i moved some default functions to the upper test case so we can call it in all tests
Changed how you end a test now you always log out at the end of a test
Changed agenda call. we now can see attendance based on status. Status has a bool if its visible in the agenda or not based on this it will show or not.
Added some default attendance statuses